requirements.txt
aiohttp==3.8.3
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
bitstring==3.1.9
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
cryptocode==0.1
cryptography==38.0.1
decorator==5.1.1
docker==6.1.3
dnspython==2.2.1
eventlet==0.33.1
executing==1.2.0
Flask==2.2.2
flask-sock==0.5.2
Flask-SocketIO==5.3.1
frozenlist==1.3.1
globals==0.3.34
google-crc32c==1.5.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
netifaces==0.10.6
numpy==1.23.4
opencv-python==4.6.0.66
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.15.0
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
python-engineio==4.3.4
python-helper==0.3.56
python-socketio==5.7.2
requests==2.28.1
simple-websocket==0.8.1
SimpleWebSocketServer==0.1.2
six==1.16.0
socketio==0.2.1
soupsieve==2.3.2.post1
stack-data==0.6.0
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
waitress==2.1.2
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
